requirements.txt
Flask>=2.0.2,<3.0.0
Flask_Cors>=3.0.10,<4.0.0
Flask_MySQLdb>=0.2.0,<1.0.0
PyJWT>=2.3.0,<3.0.0
psycopg2-binary
gunicorn
